When printing routes, note if a route traverses an unexplored sector
Because we now fill in lots of unfo for unexplored and forgotten
sectors, we will often identify routes through (or to or from) them,
but we won't know about any drones. So warn the user that the route
passes through these sectors which may (or may not) contain drones.
Also, wrap a few long lines.